# class-backend/docs/openapi.yaml
openapi: 3.1.0
info:
  title: Class Operations API
  version: 1.5.0
  license:
    name: Proprietary
    identifier: LicenseRef-ClassOps-Proprietary
  description: |
    API cho các màn React hiện có: xác thực, lập lịch lớp và vận hành giáo viên.
    JWT có `accountScope=PLATFORM|TENANT`; tenant chỉ được suy ra cho phiên TENANT.
    API tạo lớp/ca/override không nhận link online theo DEC-046.
servers:
  - url: /api/v1
security:
  - bearerAuth: []
...
paths:
...
  /schedules/management:
    get:
      operationId: getManagementSchedule
      tags: [Weekly schedules]
      summary: Lịch tuần toàn trung tâm
      parameters:
        - {$ref: "#/components/parameters/WeekStart"}
        - {name: teacherId, in: query, schema: {type: string, format: uuid}}
        - {name: roomId, in: query, schema: {type: string, format: uuid}}
      responses:
        "200":
          description: Lịch, kèm canManageSessionSchedule cho UI read-only
          content:
            application/json:
              schema: {$ref: "#/components/schemas/WeekSchedule"}
        "400": {$ref: "#/components/responses/ValidationError"}
  /schedules/me:
    get:
      operationId: getOwnSchedule
      tags: [Weekly schedules]
      summary: Lịch của giáo viên từ JWT, không nhận teacherId
      parameters:
        - {$ref: "#/components/parameters/WeekStart"}
      responses:
        "200":
          description: Lịch cá nhân
          content:
            application/json:
              schema: {$ref: "#/components/schemas/WeekSchedule"}
        "400": {$ref: "#/components/responses/ValidationError"}
  /sessions/{sessionId}/schedule-preview:
    post:
      operationId: previewSessionSchedule
      tags: [Weekly schedules]
      summary: Kiểm tra override hình thức/phòng một buổi
      parameters:
        - {$ref: "#/components/parameters/SessionId"}
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: "#/components/schemas/SessionScheduleInput"}
      responses:
        "200":
          description: Preview một buổi
          content:
            application/json:
              schema: {$ref: "#/components/schemas/SchedulePreview"}
        "409": {$ref: "#/components/responses/Conflict"}
  /sessions/{sessionId}/schedule:
    patch:
      operationId: updateSessionSchedule
      tags: [Weekly schedules]
      summary: Lưu override hình thức/phòng; không thay link online
      parameters:
        - {$ref: "#/components/parameters/SessionId"}
        - {$ref: "#/components/parameters/IdempotencyKey"}
      requestBody:
        required: true
        content:
          application/json:
            schema:
              allOf:
                - {$ref: "#/components/schemas/SessionScheduleInput"}
                - type: object
                  required: [previewId]
                  properties:
                    previewId: {type: string, format: uuid}
                    acknowledgedWarningIds:
                      type: array
                      items: {type: string}
      responses:
        "200":
          description: Lịch tuần sau cập nhật
          content:
            application/json:
              schema: {$ref: "#/components/schemas/WeekSchedule"}
        "409": {$ref: "#/components/responses/Conflict"}
  /sessions/{sessionId}/substitution-previews:
    post:
      operationId: previewSessionSubstitution
      tags: [Weekly schedules]
      summary: Kiểm tra thay giáo viên một buổi
      parameters:
        - {$ref: "#/components/parameters/SessionId"}
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: "#/components/schemas/SubstitutionPreviewInput"}
      responses:
        "200":
          description: Preview xung đột
          content: {application/json: {schema: {$ref: "#/components/schemas/SchedulePreview"}}}
        "409": {$ref: "#/components/responses/Conflict"}
  /sessions/{sessionId}/substitutions:
    post:
      operationId: applySessionSubstitution
      tags: [Weekly schedules]
      summary: Thay giáo viên thực tế của một buổi
      parameters:
        - {$ref: "#/components/parameters/SessionId"}
        - {$ref: "#/components/parameters/IdempotencyKey"}
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: "#/components/schemas/ApplySubstitutionInput"}
      responses:
        "200":
          description: Kết quả thay giáo viên
          content: {application/json: {schema: {$ref: "#/components/schemas/SessionMutationResult"}}}
        "409": {$ref: "#/components/responses/Conflict"}
  /sessions/{sessionId}/makeup-previews:
    post:
      operationId: previewSessionMakeup
      tags: [Weekly schedules]
      summary: Kiểm tra lịch buổi bù
      parameters:
        - {$ref: "#/components/parameters/SessionId"}
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: "#/components/schemas/MakeupPreviewInput"}
      responses:
        "200":
          description: Preview xung đột buổi bù
          content: {application/json: {schema: {$ref: "#/components/schemas/SchedulePreview"}}}
        "409": {$ref: "#/components/responses/Conflict"}
  /sessions/{sessionId}/cancellations:
    post:
      operationId: cancelSession
      tags: [Weekly schedules]
      summary: Hủy buổi, tùy chọn tạo buổi bù cùng transaction
      parameters:
        - {$ref: "#/components/parameters/SessionId"}
        - {$ref: "#/components/parameters/IdempotencyKey"}
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: "#/components/schemas/CancelSessionInput"}
      responses:
        "200":
          description: Buổi nguồn đã hủy và buổi bù nếu có
          content: {application/json: {schema: {$ref: "#/components/schemas/SessionMutationResult"}}}
        "409": {$ref: "#/components/responses/Conflict"}
  /sessions/{sessionId}/makeups:
    post:
      operationId: createSessionMakeup
      tags: [Weekly schedules]
      summary: Tạo buổi bù cho buổi đã hủy chưa có successor
      parameters:
        - {$ref: "#/components/parameters/SessionId"}
        - {$ref: "#/components/parameters/IdempotencyKey"}
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: "#/components/schemas/CreateMakeupInput"}
      responses:
        "200":
          description: Buổi bù mới
          content: {application/json: {schema: {$ref: "#/components/schemas/SessionMutationResult"}}}
        "409": {$ref: "#/components/responses/Conflict"}
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    SessionId: {name: sessionId, in: path, required: true, schema: {type: string, format: uuid}}
...
    WeekStart:
      name: weekStart
      in: query
      required: true
      description: Thứ Hai theo Asia/Ho_Chi_Minh
      schema: {type: string, format: date}
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema: {type: string, minLength: 1, maxLength: 150}
...
  responses:
    ValidationError:
      description: Dữ liệu không hợp lệ
      content:
        application/problem+json:
          schema: {$ref: "#/components/schemas/Problem"}
...
    Conflict:
      description: |
        Preview stale, TEACHER_OVERLAP, ROOM_OVERLAP, STUDENT_OVERLAP,
        optimistic lock hoặc idempotency conflict.
      content:
        application/problem+json:
          schema: {$ref: "#/components/schemas/Problem"}
...
  schemas:
...
    ScheduleConflict:
      type: object
      required: [id, code, severity, proposedSessionKey, studentNames, conflictingSession]
      properties:
        id: {type: string}
        code: {type: string, enum: [TEACHER_OVERLAP, ROOM_OVERLAP, STUDENT_OVERLAP]}
        severity: {type: string, enum: [BLOCKING, WARNING]}
        proposedSessionKey: {type: string}
        studentNames: {type: array, items: {type: string}}
        conflictingSession: {type: object}
    PreviewSession:
      type: object
      properties:
        key: {type: string}
        ordinal: {type: integer}
        startAt: {type: string, format: date-time}
        endAt: {type: string, format: date-time}
        teacherId: {type: string, format: uuid}
        teacherName: {type: string}
        mode: {type: string, enum: [IN_PERSON, ONLINE]}
        roomId: {type: [string, "null"], format: uuid}
        roomName: {type: [string, "null"]}
    SchedulePreview:
      type: object
      required:
        [previewId, generatedAt, expiresAt, inputVersion, sessions, skippedHolidays, expectedEndDate, conflicts]
      properties:
        previewId: {type: string, format: uuid}
        generatedAt: {type: string, format: date-time}
        expiresAt: {type: string, format: date-time}
        inputVersion: {type: string}
        sessions: {type: array, items: {$ref: "#/components/schemas/PreviewSession"}}
        skippedHolidays: {type: array, items: {type: object}}
        expectedEndDate: {type: string, format: date}
        conflicts: {type: array, items: {$ref: "#/components/schemas/ScheduleConflict"}}
    SessionScheduleInput:
      type: object
      description: Không có onlineUrl theo DEC-046.
      required: [mode, version]
      properties:
        mode: {type: string, enum: [IN_PERSON, ONLINE]}
        roomId: {type: [string, "null"], format: uuid}
        version: {type: integer, format: int64, minimum: 0}
    SubstitutionPreviewInput:
      type: object
      required: [teacherId, version]
      properties:
        teacherId: {type: string, format: uuid}
        note: {type: string}
        version: {type: integer, format: int64, minimum: 0}
    ApplySubstitutionInput:
      allOf:
        - {$ref: "#/components/schemas/SubstitutionPreviewInput"}
        - type: object
          required: [previewId]
          properties:
            previewId: {type: string, format: uuid}
            acknowledgedWarningIds: {type: array, items: {type: string}}
    MakeupScheduleInput:
      type: object
      required: [date, startTime, endTime, teacherId, mode]
      properties:
        date: {type: string, format: date}
        startTime: {type: string, format: time}
        endTime: {type: string, format: time}
        teacherId: {type: string, format: uuid}
        mode: {type: string, enum: [IN_PERSON, ONLINE]}
        roomId: {type: [string, "null"], format: uuid}
    MakeupPreviewInput:
      type: object
      required: [makeup, version]
      properties:
        makeup: {$ref: "#/components/schemas/MakeupScheduleInput"}
        version: {type: integer, format: int64, minimum: 0}
    CancelSessionInput:
      type: object
      required: [reason, version]
      properties:
        reason: {type: string, minLength: 1}
        version: {type: integer, format: int64, minimum: 0}
        makeup: {$ref: "#/components/schemas/MakeupScheduleInput"}
        previewId: {type: [string, "null"], format: uuid}
        acknowledgedWarningIds: {type: array, items: {type: string}}
    CreateMakeupInput:
      type: object
      required: [version, previewId, makeup]
      properties:
        version: {type: integer, format: int64, minimum: 0}
        previewId: {type: string, format: uuid}
        makeup: {$ref: "#/components/schemas/MakeupScheduleInput"}
        acknowledgedWarningIds: {type: array, items: {type: string}}
    SessionAction:
      type: string
      enum: [SUBSTITUTE_TEACHER, CANCEL_SESSION, CREATE_MAKEUP]
    SessionMutationView:
      type: object
      properties:
        id: {type: string, format: uuid}
        classId: {type: string, format: uuid}
        classCode: {type: string}
        className: {type: string}
        ordinal: {type: integer}
        startAt: {type: string, format: date-time}
        endAt: {type: string, format: date-time}
        plannedTeacherId: {type: string, format: uuid}
        actualTeacherId: {type: string, format: uuid}
        teacherName: {type: string}
        mode: {type: string, enum: [IN_PERSON, ONLINE]}
        roomId: {type: [string, "null"], format: uuid}
        roomName: {type: [string, "null"]}
        isSubstitution: {type: boolean}
        isMakeup: {type: boolean}
        status: {$ref: "#/components/schemas/SessionStatus"}
        makeupRootSessionId: {type: [string, "null"], format: uuid}
        replacesSessionId: {type: [string, "null"], format: uuid}
        replacementSessionId: {type: [string, "null"], format: uuid}
        cancellationReason: {type: [string, "null"]}
        allowedActions: {type: array, items: {$ref: "#/components/schemas/SessionAction"}}
        version: {type: integer, format: int64}
    SessionMutationResult:
      type: object
      properties:
        source: {$ref: "#/components/schemas/SessionMutationView"}
        makeup:
          oneOf:
            - {$ref: "#/components/schemas/SessionMutationView"}
            - type: "null"
        conflicts: {type: array, items: {$ref: "#/components/schemas/ScheduleConflict"}}
    CalendarSession:
      type: object
      properties:
        id: {type: string, format: uuid}
        classId: {type: string, format: uuid}
        classCode: {type: string}
        className: {type: string}
        ordinal: {type: integer}
        startAt: {type: string, format: date-time}
        endAt: {type: string, format: date-time}
        plannedTeacherId: {type: string, format: uuid}
        actualTeacherId: {type: string, format: uuid}
        teacherName: {type: string}
        mode: {type: string, enum: [IN_PERSON, ONLINE]}
        roomId: {type: [string, "null"], format: uuid}
        roomName: {type: [string, "null"]}
        onlineUrl:
          type: string
          readOnly: true
          description: Link do check-in tương lai thiết lập; scheduling API không ghi.
        isSubstitution: {type: boolean}
        isMakeup: {type: boolean}
        status: {$ref: "#/components/schemas/SessionStatus"}
        scheduleState:
          type: string
          enum: [UPCOMING, TAUGHT, MISSING_CHECK_IN, CANCELLED]
          description: Trạng thái hiển thị trên thời khóa biểu, tính theo giờ máy chủ và dữ liệu check-in.
        makeupRootSessionId: {type: [string, "null"], format: uuid}
        replacesSessionId: {type: [string, "null"], format: uuid}
        replacementSessionId: {type: [string, "null"], format: uuid}
        cancellationReason: {type: [string, "null"]}
        allowedActions: {type: array, items: {$ref: "#/components/schemas/SessionAction"}}
        version: {type: integer, format: int64}
    WeekSchedule:
      type: object
      properties:
        weekStart: {type: string, format: date}
        weekEnd: {type: string, format: date}
        sessions: {type: array, items: {$ref: "#/components/schemas/CalendarSession"}}
        canManageSessionSchedule: {type: boolean}
...
    SessionStatus:
      type: string
      enum: [SCHEDULED, IN_PROGRESS, COMPLETED, PENDING_CONFIRMATION, CANCELLED]
...
    Problem:
      type: object
      required: [type, title, status, detail, code, message, timestamp, traceId]
      properties:
        type: {type: string, format: uri}
        title: {type: string}
        status: {type: integer}
        detail: {type: string}
        instance: {type: string}
        code: {type: string}
        message: {type: string}
        fieldErrors: {type: object, additionalProperties: {type: string}}
        details: {type: object, additionalProperties: true}
        retryAfterSeconds: {type: integer, minimum: 1}
        timestamp: {type: string, format: date-time}
        traceId: {type: string}